+++ app/src/components/redlist/RedListView.tsx
@@ -3751,6 +3751,7 @@ export default function RedListView({ viewMode = "reassessments", sharedTaxa, sh
                                 assessmentYear={assessmentYear}
                                 assessmentDate={s.assessment_date}
                                 taxonGroup={s.taxon_group}
+                                nativeCountries={s.countries}
                                 onEmpty={s.category === "NE" ? handleOccurrenceEmpty : undefined}
                               />
                             </div>